feat(front): refactors ThemeToggle component
