+++ front/src/components/ThemeToggle.svelte
@@ -1,47 +1,46 @@
 <script>
-  export let dark = null;
+  import { onMount } from 'svelte';
+  import ThemeToggleIcon from "./ThemeToggleIcon.svelte";
+
+  const light = "light-theme";
+  let root;
+  let body;
+  let darkTheme = "dark-theme";
+  let lightTheme = "light-theme";
+  $: theme = "";
+  $: loaded = false;
+  $: dark = null;
+
+  onMount(async () => {
+    loaded = true;
+    const themeListener = window.matchMedia('(prefers-color-scheme: dark)');
+    window.matchMedia && themeListener.matches
+    ? dark = true
+    : dark = false;
+    theme = dark ? darkTheme : lightTheme;
+    root = document.getElementsByTagName("html")[0];
+    body = document.body;
+
+    root.classList = `${theme}`;
+    body.classList.add("ready");
+
+    window.matchMedia('(prefers-color-scheme: dark)')
+      .addEventListener('change', () => toggleTheme());
+  });
+
+  function toggleTheme() {
+    dark = !dark;
+    root.classList.toggle(darkTheme);
+    root.classList.toggle(lightTheme);
+    body.classList.toggle(darkTheme);
+    body.classList.toggle(lightTheme);
+  }
+
+  const baseClasses = "theme-toggle"
+  $: classList = `${baseClasses} ${loaded ? "opacity-100" : "opacity-0"}`;
+
 </script>
 
-{#if dark === true}
-  <span class="sr-only">Turn on light theme.</span>
-  <svg
-    id="sun"
-    xmlns="http://www.w3.org/2000/svg"
-    width="20"
-    height="20"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round"
-    aria-hidden="true"
-  >
-    <circle cx="12" cy="12" r="5" />
-    <line x1="12" y1="1" x2="12" y2="3" />
-    <line x1="12" y1="21" x2="12" y2="23" />
-    <line x1="4.22" y1="4.22" x2="5.64" y2="5.64" />
-    <line x1="18.36" y1="18.36" x2="19.78" y2="19.78" />
-    <line x1="1" y1="12" x2="3" y2="12" />
-    <line x1="21" y1="12" x2="23" y2="12" />
-    <line x1="4.22" y1="19.78" x2="5.64" y2="18.36" />
-    <line x1="18.36" y1="5.64" x2="19.78" y2="4.22" />
-  </svg>
-{:else}
-  <span class="sr-only">Turn on dark theme.</span>
-  <svg
-    id="moon"
-    xmlns="http://www.w3.org/2000/svg"
-    width="20"
-    height="20"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round"
-    aria-hidden="true"
-  >
-    <path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z" />
-  </svg>
-{/if}
+<button class={classList} on:click={toggleTheme}>
+  <ThemeToggleIcon {dark} />
+</button>
